backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
torch==2.6.0
torchvision==0.21.0
torchaudio==2.6.0
transformers==4.37.2
opencv-python==4.9.0.80
numpy==1.26.3
PyMuPDF==1.23.8
PyPDF2==3.0.1
pytesseract==0.3.10
Pillow==10.2.0
python-dotenv==1.0.1
groq==0.4.2
moviepy==1.0.3
imageio-ffmpeg==0.4.9 